requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
django-autoslug==1.9.9
django-cors-headers==4.0.0
django-countries==7.5.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.11
Pillow==9.5.0
pkg_resources==0.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
urllib3==2.0.2
